pyproject.toml
[tool.poetry]
name = "achemi"
version = "0.1.0"
description = ""
authors = ["mihasim <misha.dorozhkin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.27"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
pytelegrambotapi = "^4.16.1"
schedule = "^1.2.1"
apscheduler = "^3.10.4"
pytest = "^8.0.1"
coverage = "^7.4.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
